+++ features/steps/flaw_create.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from behave import *
 
 from features.pages.advanced_search_page import AdvancedSearchPage
-from features.pages.flaw_detail_page import FlawDetailPage
+from features.pages.flaw_create_page import FlawCreatePage
 from features.utils import (
     generate_cve,
     generate_cwe,
@@ -16,33 +16,32 @@ MAX_RETRY = 10
 
 
 def create_flaw_with_valid_data(context, embargoed=False, with_optional=False):
-    flaw_page = FlawDetailPage(context.browser)
+    flaw_create_page = FlawCreatePage(context.browser)
     context.cve_id = generate_cve()
-    description = generate_random_text()
     reported_date = datetime.today().strftime("%Y%m%d")
-    flaw_page.set_input_field('title', generate_random_text())
-    flaw_page.set_input_field('component', 'autocomponent')
-    flaw_page.set_input_field('cveid', context.cve_id)
-    flaw_page.set_input_field('reportedDate', reported_date)
-    flaw_page.set_select_value('source')
-    flaw_page.set_select_value('impact')
+    flaw_create_page.set_input_field('title', generate_random_text())
+    flaw_create_page.set_input_field('component', 'autocomponent')
+    flaw_create_page.set_input_field('cveid', context.cve_id)
+    flaw_create_page.set_input_field('reportedDate', reported_date)
+    flaw_create_page.set_select_value('impact')
+    flaw_create_page.set_select_value('source')
     if embargoed:
-        flaw_page.click_btn("embargeodCheckBox")
+        flaw_create_page.click_btn("embargeodCheckBox")
     else:
-        flaw_page.set_input_field('publicDate', reported_date)
-    flaw_page.set_document_text_field('comment#0', generate_random_text())
+        flaw_create_page.set_input_field('publicDate', reported_date)
+    flaw_create_page.set_document_text_field('comment#0', generate_random_text())
     if with_optional:
-        flaw_page.set_input_field('cweid', generate_cwe())
-        flaw_page.set_document_text_field('description', generate_random_text())
-        flaw_page.set_document_text_field('statement', generate_random_text())
+        flaw_create_page.set_input_field('cweid', generate_cwe())
+        flaw_create_page.set_document_text_field('description', generate_random_text())
+        flaw_create_page.set_document_text_field('statement', generate_random_text())
     count = 0
     while count < MAX_RETRY:
-        flaw_page.click_btn('createNewFlawBtn')
+        flaw_create_page.click_btn('createNewFlawBtn')
         try:
-            flaw_page.wait_msg('flawCreatedMsg')
+            flaw_create_page.wait_msg('flawCreatedMsg')
         except Exception:
             context.cve_id = generate_cve()
-            flaw_page.set_input_field('cveid', context.cve_id)
+            flaw_create_page.set_input_field('cveid', context.cve_id)
             count += 1
         else:
             break
@@ -60,7 +59,7 @@ def check_created_flaw_exist(context, embargoed=False):
 
 @given('I open the flaw create page')
 def step_impl(context):
-    flaw_page = FlawDetailPage(context.browser)
+    flaw_page = FlawCreatePage(context.browser)
     flaw_page.click_btn('createFlawLink')