+++ src/components/IssueQueue.vue
@@ -14,7 +14,10 @@ type FilteredIssue = {
   selected: boolean;
 };
 
-type ColumnField = 'id' | 'impact' | 'source' | 'created_dt' | 'title' | 'state' | 'owner';
+// Temporarily hiding 'Source' column to avoid displaying incorrect information.
+// TODO: unhide it once final issue sources are defined. [OSIDB-2424]
+// type ColumnField = 'id' | 'impact' | 'source' | 'created_dt' | 'title' | 'state' | 'owner'; 
+type ColumnField = 'id' | 'impact' | 'created_dt' | 'title' | 'state' | 'owner';
 
 const props = defineProps<{
   issues: any[];
@@ -64,7 +67,7 @@ const params = computed(() => {
 const columnsFieldsMap: Record<string, ColumnField> = {
   ID: 'id',
   Impact: 'impact',
-  Source: 'source',
+  // Source: 'source',
   Created: 'created_dt',
   Title: 'title',
   State: 'state',
@@ -103,7 +106,7 @@ function relevantFields(issue: any) {
   return {
     id: issue.cve_id || issue.uuid,
     impact: issue.impact,
-    source: issue.source,
+    // source: issue.source,
     created_dt: issue.created_dt,
     title: issue.title,
     workflowState: issue.classification.state,
@@ -244,7 +247,7 @@ watch(params, () => {
   div.osim-incident-list {
     display: block;
     max-height: calc(100vh - 164px);
-    overflow-y: scroll;
+    overflow-y: auto;
 
     &:hover::-webkit-scrollbar-thumb {
       background-color: $redhat-red-50;