+++ CHANGELOG.md
@@ -16,7 +16,10 @@
 * Provided save operation indications
 * Implemented read-only mode (network requests involving write operations are disabled)
 * Disables form on save
+* Links on references are now displayed as urls
 * Implemented Advanced Search Query to URL
+* `Source` column hidden on flaw lists
+* `Source` field renamed to `CVE Source` on flaw form and advanced search
 * Added Total count for IndexView and AdvancedSearch Page
 
 ### Fixed
@@ -38,6 +41,9 @@
 * Fixed `Reference` external url validation to accept `http://`
 * Date Field: Restore clickability to save/cancel buttons on Date field, Fixes escape key functionality, Prevents reappearing value from cancelled edit
 * Fixed NVD and RH CVSS Score diff highlight
+* Fixed duplicated search option on `AdvancedSearch` page
+* Fixed redirect `Flaws` link on `Flaw` edit page when `Flaw` is not found.
+* Fixed Scroll bar on `Flaws` list
 
 
 ## [2024.1.0]